+++ onlinenewssite/newssubscriber/z/index.php
@@ -10,7 +10,7 @@
  * @copyright 2018 Hardcover LLC
  * @license   https://hardcoverwebdesign.com/license  MIT License
  *            https://hardcoverwebdesign.com/gpl-2.0  GNU General Public License, Version 2
- * @version:  2018 05 13
+ * @version:  2018 09 28
  * @link      https://hardcoverwebdesign.com/
  * @link      https://online-news-site.com/
  * @link      https://github.com/hardcover/
@@ -129,14 +129,46 @@ if ($task == 'adSync') {
 // Archives
 //
 if ($task == 'archiveDelete') {
-    $dbh = new PDO($dbArchive);
-    $stmt = $dbh->prepare('DELETE FROM articles WHERE rowid=?');
-    $stmt->execute([$idArticle]);
-    $dbh = null;
-    $dbh = new PDO($dbArchive2);
-    $stmt = $dbh->prepare('DELETE FROM imageSecondary WHERE idArticle=?');
-    $stmt->execute([$idArticle]);
-    $dbh = null;
+    $dbNumber = 0;
+    while ($dbNumber !== -1) {
+        $db = str_replace('archive', 'archive-' . $dbNumber, $dbArchive);
+        if ($dbNumber === 0
+            or file_exists(str_replace('sqlite:', '', $db))
+        ) {
+            if ($dbNumber === 0) {
+                $database = $dbArchive;
+            } else {
+                $database = $db;
+            }
+            $dbNumber++;
+        } else {
+            $dbNumber = -1;
+            $dbh = null;
+        }
+        if ($database != null) {
+            $dbh = new PDO($database);
+            $stmt = $dbh->prepare('SELECT idArticle FROM articles WHERE idArticle=?');
+            $stmt->setFetchMode(PDO::FETCH_ASSOC);
+            $stmt->execute([$idArticle]);
+            $row = $stmt->fetch();
+            if ($row) {
+                $stmt = $dbh->prepare('DELETE FROM articles WHERE idArticle=?');
+                $stmt->execute([$idArticle]);
+                $dbh = null;
+                $db = str_replace('archive2', 'archive2-' . $dbNumber, $dbArchive2);
+                if ($dbNumber === 0) {
+                    $database = $dbArchive2;
+                } else {
+                    $database = $db;
+                }
+                $dbh = new PDO($database);
+                $stmt = $dbh->prepare('DELETE FROM imageSecondary WHERE idArticle=?');
+                $stmt->execute([$idArticle]);
+                $dbh = null;
+                $dbNumber = -1;
+            }
+        }
+    }
     $response['result'] = 'success';
 }
 if ($task == 'archiveSearch') {
